# Speed up consul services deregistration during sync