New | Loops | Speed [ns/op]
------------ | -------------|--------
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps10x2-4   |	  300000	|      4684
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x2-4  |	   30000	|     53626
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x100-4	|     300 |	   5359705

Old | Loops | Speed [ns/op]
------------ | -------------|--------
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps10x2-4   |	  200000	   |   6307
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x2-4  |	    2000	|    608956
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x100-4	 |      1 |	1460033471

Generated with `go test ./sync -bench=.`