@dankraw 2015-12-16
Just tried it on my machine, looks even better. :boom: 

New | Loops | Speed [ns/op]
------------ | ------------- | --------
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps10x2-4 | 500000 | 2897 ns/op
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x2-4 | 50000 | 34168 ns/op
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x100-4 | 500 | 3350642 ns/op