@wendigo 2015-12-16
New | Loops | Speed [ns/op]
------------ | ------------- | ------------
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps10x2-8   	 | 500000	 |     2739 ns/op
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x2-8  	 |  50000	  |   33058 ns/op
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x100-8 |	     500	|   2941646 ns/op

Yep, definitely better :dancer: