@janisz 2015-12-16
How about previous version? It's no surprise it's faster on faster machine.