@dankraw 2015-12-17
Noticeably worse:

Old | Loops | Speed [ns/op]
---- | --- | ----
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps10x2-4       |    300000    |          5458
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x2-4      |      3000   |         515053
BenchmarkDeregisterConsulServicesThatAreNotInMarathonApps100x100-4   |          1  |      1279499301